move report endpoint + test suite into posthog/api/report.py and adjust router etc.
